e2e/serverのテストでなぜcreateCandidateBufferが必要になったのでしょうか？WebRTCの仕様的に本来こうする必要がないのであれば実装側にデグレが発生した可能性があるので原因を調査して修正してください
